Move email.py to helpers
